% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadkml.R
\name{loadkml}
\alias{loadkml}
\title{loadkml}
\usage{
loadkml(folder = NULL, layer = "track_points")
}
\arguments{
\item{folder}{select the folder containing the kml file(s)}

\item{layer}{select layer to use for reading the data}
}
\value{
kml tidy data frame containing loaded data
}
\description{
Load kml files and return a tidy data frame (tested with files generated by KDE Marble).
}

