% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadmeerun.R
\name{loadmeerun}
\alias{loadmeerun}
\title{loadmeerun}
\usage{
loadmeerun(folder = NULL, layer = "track_points")
}
\arguments{
\item{folder}{select the folder containing the kml file(s)}

\item{layer}{select layer to use for reading the data}
}
\value{
gpx tidy data frame containing loaded data
}
\description{
Load gpx files and return a tidy data frame (tested with files generated by MeeRun).
}

