% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadrunkeeper.R
\name{loadrunkeeper}
\alias{loadrunkeeper}
\title{loadrunkeeper}
\usage{
loadrunkeeper(folder)
}
\arguments{
\item{folder}{select the folder containing the kml file(s)}
}
\value{
gpx tidy data frame containing loaded data
}
\description{
Load gpx files and return a tidy data frame (function adapted from FlowingData).
}

