% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgpx.R
\name{plotgpx}
\alias{plotgpx}
\title{plotgpx}
\usage{
plotgpx(folder, type = "osm", zoom = 14, col = "purple",
  mapborder = 0.1, pdftitle = "Map plot")
}
\arguments{
\item{folder}{select the folder containing the kml file(s)}

\item{type}{the type of map}

\item{zoom}{zoom level of the map}

\item{col}{colour of the GPS tracks}

\item{mapborder}{ratio by which the height and width of the map are increased}

\item{pdftitle}{title of the resulting pdf file}
}
\value{
kml tidy data frame containing loaded data
}
\description{
Create a map with GPS tracks using map parameters extracted from used tracks
}
\references{
help and infos from:
- http://dbsgeo.com/latlon/
- http://www.r-bloggers.com/plot-maps-like-a-boss/
- http://stackoverflow.com/questions/13762793/plotting-choropleth-maps-from-kml-data-using-ggplot2
- http://flowingdata.com/2011/05/11/how-to-map-connections-with-great-circles/
- http://rpsychologist.com/parsing-data-from-a-text-file-and-plotting-where-people-live-using-ggplot2-and-openstreetmaps
}

